refactor: extract shared primitives into _core.py (#98 Phase C, step 1)
First half of the annotation-pass split. Before the passes can move into their
own module, the primitives they share with the rest of make_drawing.py must
live below both, or the new module would import back from make_drawing.py and
recreate a cycle.

Move the shared data structures, helpers, and layout constants into a new
`draftwright/_core.py` that imports only from `draftwright.layout` and
third-party libs — never from make_drawing.py — keeping the module graph a DAG
(layout → _core → make_drawing), consistent with how layout.py/pmi.py already
sit below the domain API.

Moved: `Analysis`/`_Projector`/`Strip`/`ViewZones`; `_dim`, `_fmt`,
`_axis_letter`, `_tag_sequence`, `_legible_steps`, `_largest_empty_rect`,
`_iso_bbox`, `_add_title_block`; and the page/slot/margin constants plus `_log`
and the strip-solver aliases. make_drawing.py imports them back (the 26 still
referenced there); two constants used only by moved code (`_PAD`,
`_MIN_STEP_DIM_MM`) are not re-imported. Three test imports of those two
constants are repointed to `_core`.

Pure move — no behaviour change. ruff/mypy clean; make_drawing tests pass.
